feat: Closes #35,  fetch posts from meumobi API
