+++ src/pages/post-details/post-details.ts
@@ -2,7 +2,7 @@ import { Component, ViewChild } from '@angular/core';
 import { IonicPage, NavController, NavParams } from 'ionic-angular';
 
 import { Post } from '@models/post.interface';
-import { PostsProvider } from '@providers/posts';
+import { PostsService } from '@providers/posts';
 import { MeuToastProvider } from '@shared/meu-toast.service';
 
 @IonicPage({
@@ -20,7 +20,7 @@ export class PostDetailsPage {
   post: Post;
 
   constructor(
-    private postsProvider: PostsProvider,
+    private postsService: PostsService,
     public rootNavCtrl: NavController, 
     public navParams: NavParams,
     public toast: MeuToastProvider,
@@ -32,8 +32,10 @@ export class PostDetailsPage {
   }
   
   findById(id) {
-    this.postsProvider.findById(id)
+    console.log(id);
+    this.postsService.findById(id)
       .then(data => {
+        console.log(data);
         this.post = data;
       })
   }
