+++ src/pages/posts/posts.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 
 import { PostsPage } from './posts';
-import { PostsProvider } from '@providers/posts';
+import { PostsService } from '@providers/posts';
 
 @NgModule({
   declarations: [
@@ -12,7 +12,7 @@ import { PostsProvider } from '@providers/posts';
     IonicPageModule.forChild(PostsPage),
   ],
   providers: [
-    PostsProvider
+    PostsService
   ],
   entryComponents: [
   	PostsPage
